requirements.txt
discord.py==1.5.0
bs4~=0.0.1

requests~=2.24.0
beautifulsoup4~=4.9.3